% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_normal.R
\name{generate_X}
\alias{generate_X}
\title{Simulate a normal data set \eqn{S = (X)}.}
\usage{
generate_X(n = 100, mu = rep(0, 10), sigma = rep(1, 10))
}
\arguments{
\item{n}{The desired number of data points in the data set.}

\item{mu}{A \eqn{p}-dimensional vector of means  for \eqn{\mu}.}

\item{sigma}{A \eqn{p}-dimensional vector of non-negative standard deviations for \eqn{\sigma}.}
}
\value{
An \eqn{n \times p} dimensional data frame given by \eqn{S = (X)}.
In the default case, the columns \eqn{X_i} are sampled from \eqn{N(0,1)}, \eqn{n = 100} and \eqn{p = 10}.
}
\description{
Creates a toy data set \eqn{S = (X)} where the columns of \eqn{X}
are sampled from an independent Gaussian distribution with mean \eqn{\mu_i} and
standard deviation \eqn{\sigma_i}, i.e. \eqn{N(\mu_i, \sigma_i^2)}.
The final dimension will be \eqn{n \times p},
with the number of data points \eqn{n} to be specified.
}
\examples{
generate_X()

generate_X(n = 40, mu = 1:10, sigma = rep(1, 10))
}
