% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_normal.R
\name{generate_X_cat}
\alias{generate_X_cat}
\title{Simulating a normal data set \eqn{S} that includes categorical variables.}
\usage{
generate_X_cat(
  n = 100,
  mu = rep(0, 10),
  sigma = rep(1, 10),
  no_of_cat = c(4, 5)
)
}
\arguments{
\item{n}{desired number of data points in the data set.}

\item{mu}{a \eqn{p_1}-dimensional vector of means for \eqn{\mu}.}

\item{sigma}{a \eqn{p_1}-dimensional vector of non-negative standard deviations for \eqn{\sigma}.}

\item{no_of_cat}{a \eqn{p_2}-dimensional vector where the entries indicate the number of categories desired for each column of \eqn{X_2}.}
}
\value{
An \eqn{n \times p} dimensional data frame given by \eqn{S = (X_1, X_2)}.
In the base case, the columns of \eqn{X_1} are sampled from \eqn{N(0,1)}, \eqn{n = 100} and \eqn{p = 10},
and two additional categorical columns of \eqn{X_2} are added.
}
\description{
Creates a data set \eqn{S = (X_1, X_2)} where the columns of \eqn{X_1}
are sampled from an independent Gaussian distribution with mean \eqn{\mu_i} and
standard deviation \eqn{\sigma_i}, i.e. \eqn{N(\mu_i, \sigma_i^2)},
and the columns of \eqn{X_2} are categorical, sampled with replacement from a given number of categories.
The final dimension will be \eqn{n \times (p_1 + p_2)},
where \eqn{p_1} is the number of columns in \eqn{X_1} and \eqn{p_2} is the number of columns in \eqn{X_2},
with the number of data points \eqn{n} to be specified.
}
\examples{
generate_X_cat()
generate_X_cat(n = 40, mu = 1:10, sigma = rep(1, 10), no_of_cat = c(2,3))
}
